Incorporate executor-queue into frontend server
