--- enterprise/cmd/frontend/internal/executor/CODENOTIFY
@@ -1,3 +1,0 @@
-# See https://github.com/sourcegraph/codenotify for documentation.
-
-**/* @efritz
