+++ enterprise/cmd/frontend/internal/executorqueue/CODENOTIFY
@@ -1,3 +1,4 @@
 # See https://github.com/sourcegraph/codenotify for documentation.
 
 **/* @efritz
+**/* @eseliger
