+++ enterprise/cmd/frontend/internal/executorqueue/config/shared.go
@@ -1,6 +1,10 @@
 package config
 
-import "github.com/sourcegraph/sourcegraph/internal/env"
+import (
+	"github.com/cockroachdb/errors"
+
+	"github.com/sourcegraph/sourcegraph/internal/env"
+)
 
 // SharedConfig defines common items that are used by multiple queues.
 type SharedConfig struct {
@@ -15,4 +19,14 @@ func (c *SharedConfig) Load() {
 	c.FrontendURL = c.Get("EXECUTOR_FRONTEND_URL", "", "The external URL of the sourcegraph instance.")
 	c.FrontendUsername = c.Get("EXECUTOR_FRONTEND_USERNAME", "", "The username supplied to the frontend.")
 	c.FrontendPassword = c.Get("EXECUTOR_FRONTEND_PASSWORD", "", "The password supplied to the frontend.")
+}
+
+func (c *SharedConfig) Validate() error {
+	if c.FrontendUsername == "" {
+		return errors.Errorf("invalid value for EXECUTOR_FRONTEND_USERNAME: no value supplied")
+	}
+	if c.FrontendPassword == "" {
+		return errors.Errorf("invalid value for EXECUTOR_FRONTEND_PASSWORD: no value supplied")
+	}
+	return c.BaseConfig.Validate()
 }
