+++ enterprise/cmd/frontend/internal/executorqueue/handler/handler.go
@@ -1,4 +1,4 @@
-package server
+package handler
 
 import (
 	"context"
