+++ enterprise/cmd/frontend/internal/executorqueue/handler/handler_test.go
@@ -1,4 +1,4 @@
-package server
+package handler
 
 import (
 	"context"
