+++ enterprise/cmd/frontend/internal/executorqueue/queues/batches/config.go
@@ -1,7 +1,7 @@
 package batches
 
 import (
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/internal/config"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executorqueue/config"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 )
