+++ enterprise/cmd/frontend/internal/executorqueue/queues/codeintel/config.go
@@ -1,7 +1,7 @@
 package codeintel
 
 import (
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/internal/config"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executorqueue/config"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 )
